Revert "Basic architecture of Maintenance App - SUST-19, SUST-42. Implement Force Publish Course view. SUST-46"
This reverts commit 27646b7e6fa374f1b48b29896fb2a24ef0d9b4e7.
