+++ cms/envs/common.py
@@ -828,9 +828,6 @@ INSTALLED_APPS = (
     'openedx.core.djangoapps.coursetalk',  # not used in cms (yet), but tests run
     'xblock_config',
 
-    # Maintenance tools
-    'maintenance',
-
     # Tracking
     'track',
     'eventtracking.django.apps.EventTrackingConfig',
