+++ cms/static/sass/_build-v1.scss
@@ -68,7 +68,6 @@
 @import 'views/group-configuration';
 @import 'views/video-upload';
 @import 'views/certificates';
-@import 'views/maintenance';
 
 // +Base - Contexts
 // ====================
