--- cms/static/sass/views/_maintenance.scss
@@ -1,71 +1,0 @@
-.maintenance-header {
-    text-align: center;
-    margin-top: 50px;
-
-    h2 {
-      margin-bottom: 10px;
-    }
-}
-.maintenance-content {
-  padding: 3rem 0;
-  .maintenance-list {
-    max-width: 1280px;
-    margin: 0 auto;
-    .view-list-container {
-      padding: 10px 15px;
-      background-color: #fff;
-      border-bottom: 1px solid #ddd;
-      &:hover {
-        background-color: #fafafa;
-      }
-      .view-name {
-        display: inline-block;
-        width: 20%;
-        float: left;
-      }
-      .view-desc {
-        display: inline-block;
-        width: 80%;
-      font-size: 15px;
-      }
-    }
-  }
-  .maintenance-form {
-    width: 60%;
-    margin: auto;
-    .result-list {
-      height: calc(100vh - 200px);
-      overflow: auto;
-    }
-    .result{
-      box-shadow: inset 0 1px 2px rgba(0, 0, 0, 0.2);
-      margin-top: 15px;
-      padding: 15px 30px;
-      background: #f9f9f9;
-    }
-    li {
-      font-size: 13px;
-      line-height: 9px;
-    }
-    .actions {
-      text-align: right;
-    }
-    .field-radio div {
-      display: inline-block;
-      margin-right: 10px;
-    }
-    div.error {
-      color: #F00;
-      margin-top: 10px;
-      font-size: 13px;
-    }
-    div.head-output {
-      font-size: 13px;
-      margin-bottom: 10px;
-    }
-    div.main-output {
-      color: #0A0;
-      font-size: 15px;
-    }
-  }
-}
