--- cms/templates/js/maintenance/force-published-course-response.underscore
@@ -1,14 +1,0 @@
-<div class="result">
-    <div class="head-output">
-        <%- gettext('You have done a dry run of force publishing the course. Nothing has changed. Had you run it, the following course versions would have been change.') %>
-    </div>
-    <div class="main-output">
-        <%= StringUtils.interpolate(
-            gettext('The published branch version, {published}, was reset to the draft branch version, {draft}.'),
-            {
-                published: current_versions['published-branch'],
-                draft: current_versions['draft-branch']
-            })
-        %>
-    </div>
-</div>
