--- cms/templates/maintenance/_force_publish_course.html
@@ -1,33 +1,0 @@
-<%page expression_filter="h"/>
-<%namespace name='static' file='../static_content.html'/>
-<%!
-from django.utils.translation import ugettext as _
-from openedx.core.djangolib.markup import HTML, Text
-%>
-<div id="force-published-form" class="wrap-instructor-info studio-view maintenance-form">
-    <form id="force_publish" class="form-create" method="post">
-        <input type="hidden" name="csrfmiddlewaretoken" value="${csrf_token}"/>
-        <div class="wrapper-form">
-            <fieldset>
-                <legend class="sr">${_("Required data to force publish course.")}</legend>
-                <div class="list-input">
-                    <div id="course-id-container" class="field text required">
-                        <label for="course-id">${_('Course ID')}</label>
-                        <input id="course-id" type="text" name="course-id" aria-describedby="course-id-desc" required />
-                        <div id="course-id-desc" class="tip tip-stacked">${_('course-v1:edX+DemoX+Demo_Course')}</div>
-                        <div class="wrapper-error"></div>
-                    </div>
-                </div>
-            </fieldset>
-        </div>
-        <div class="actions">
-            <button type="submit" id="submit_force_publish" class="action action-primary">${_('Force Publish Course')}
-            </button>
-            <button id="reset-button" class="action action-secondary action-cancel"
-               aria-describedby="reset-values-desc">${_('Reset')}</button>
-            <span id="reset-values-desc" class="is-hidden">${_('Reset values')}</span>
-        </div>
-    </form>
-    <div id="result-error"><div class="wrapper-error"></div></div>
-    <div id="result-container" class="result-container"></div>
-</div>
