--- cms/templates/maintenance/base.html
@@ -1,21 +1,0 @@
-<%page expression_filter="h"/>
-<%inherit file="../base.html" />
-<%def name='online_help_token()'><% return 'maintenance' %></%def>
-<%namespace name='static' file='../static_content.html'/>
-<%!
-from django.core.urlresolvers import reverse
-from django.utils.translation import ugettext as _
-%>
-<%block name="content">
-<div class="wrapper-content wrapper">
-    <div class="maintenance-header">
-        <h2>
-            <a href="${reverse('maintenance:maintenance_index')}">
-                <span>${_('Maintenance Dashboard')}</span>
-            </a>
-        </h2>
-        <%block name="viewtitle">
-    </%block>
-</div>
-<%block name="viewcontent"></%block>
-</%block>
