--- cms/templates/maintenance/container.html
@@ -1,28 +1,0 @@
-<%page expression_filter="h"/>
-<%inherit file="base.html" />
-<%namespace name='static' file='../static_content.html'/>
-<%!
-from django.core.urlresolvers import reverse
-from openedx.core.djangolib.js_utils import js_escaped_string
-%>
-<%block name="title">${view['name']}</%block>
-<%block name="viewtitle">
-<h3 class="info-course">
-    <span>${view['name']}</span>
-</h3>
-</%block>
-
-<%block name="js_extra">
-   <script src="${static.url('js/maintenance/force_publish.js')}"></script>
-</%block>
-
-<%block name="viewcontent">
-<section class="container maintenance-content">
-    <%include file="_${view['slug']}.html"/>
-    <%block name="requirejs">
-      require(["js/maintenance/${view['slug'] | n, js_escaped_string}"], function(MaintenanceFactory) {
-          MaintenanceFactory("${reverse(view['url']) | n, js_escaped_string}");
-      });
-    </%block>
-</section>
-</%block>
