--- cms/templates/maintenance/index.html
@@ -1,20 +1,0 @@
-<%page expression_filter="h"/>
-<%inherit file="base.html" />
-<%namespace name='static' file='../static_content.html'/>
-<%!
-from django.utils.translation import ugettext as _
-from django.core.urlresolvers import reverse
-%>
-<%block name="title">${_('Maintenance Dashboard')}</%block>
-<%block name="viewcontent">
-<div class="container maintenance-content">
-    <ul class="maintenance-list">
-        % for view in views.values():
-        <li class="view-list-container">
-            <a class="view-name" href='${reverse(view["url"])}'>${view['name']}</a>
-            <span class="view-desc">${view['description']}</span>
-        </li>
-        % endfor
-    </ul>
-</div>
-</%block>
