+++ cms/templates/widgets/user_dropdown.html
@@ -4,7 +4,6 @@
   from django.conf import settings
   from django.core.urlresolvers import reverse
   from django.utils.translation import ugettext as _
-  from student.roles import GlobalStaff
 %>
 
 % if uses_pattern_library:
@@ -46,11 +45,6 @@
         <li class="nav-item nav-account-dashboard">
           <a href="/">${_("{studio_name} Home").format(studio_name=settings.STUDIO_SHORT_NAME)}</a>
         </li>
-        % if GlobalStaff().has_user(user):
-        <li class="nav-item">
-          <a href="${reverse('maintenance:maintenance_index')}">${_("Maintenance")}</a>
-        </li>
-        % endif
         <li class="nav-item nav-account-signout">
           <a class="action action-signout" href="${reverse('logout')}">${_("Sign Out")}</a>
         </li>
